+++ pkg/data/client/client.go
@@ -4,12 +4,10 @@ import (
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	backupv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/backup/v1alpha1"
 	corev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/core/v1alpha1"
-	examplev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/example/v1alpha1"
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	providerv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/provider/v1alpha1"
 	releasev1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/release/v1alpha1"
 	securityv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
-	toolingv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/tooling/v1alpha1"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
@@ -49,12 +47,10 @@ func New(config Config) (*Client, error) {
 				applicationv1alpha1.AddToScheme,
 				backupv1alpha1.AddToScheme,
 				corev1alpha1.AddToScheme,
-				examplev1alpha1.AddToScheme,
 				infrastructurev1alpha3.AddToScheme,
 				providerv1alpha1.AddToScheme,
 				releasev1alpha1.AddToScheme,
 				securityv1alpha1.AddToScheme,
-				toolingv1alpha1.AddToScheme,
 			},
 			Logger:     config.Logger,
 			RestConfig: config.K8sRestConfig,